+++ zigzag.legendweathers/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+- Added the Faceless model version (can be activated in the config)
+- Improved the Majora Moon position in the sky for the following : Moons of Otherworldly Oddity, Nightmare Moons, Harvest Moons, Kast, Ganimedes, Sector-0, Secret Labs, Orion, Aquatis, Wither
+
 ## 1.0.1
 - Fixed texture on the Majora Head for the N64 model version
 - Fixed Masked enemies invocation having the wrong player suit when playing in singleplayer